django-backend/requirements.txt
asgiref==3.9.1
dj-database-url==3.0.1
Django==5.2.5
django-cors-headers==4.7.0
djangorestframework==3.16.1
gunicorn==23.0.0
packaging==25.0
psycopg2-binary==2.9.10
sqlparse==0.5.3
whitenoise==6.9.0
djangorestframework-simplejwt

# For Celery and Redis
